Renamed convolution_gridding.py to gridding.py
